Make commit message field types not-generic
Summary:
Previous diffs made the types for commit message fields be generic and statically typed, differing between internal and external. We're instead going to go for a completely dynamic field config determined by sl config. This means we should make these types not statically known at all. This diff is the first step in doing this.

Next, we can remove the internal only implementations of utils and make the field config be returned by the server.

Reviewed By: muirdm

Differential Revision: D44688588

fbshipit-source-id: e63d6343596e48a0563975763c438e45d615f5b4