Make commitMessageFieldsSchema into a recoil atom
Summary:
The field config schema defines which fields we expect in commit messages in this repo, and how the user should interact with them.
Previously, we had the field config schema defined as a global constant, overwritten for internal builds.

instead, we want this to be a fully dynamic value, which could be set by an sl config. This allows repos to configure their own commit message field formats, in case you want to always have a summary/test plan/fixes PR #xyz, etc.

Making this get set in recoil means we can dynamically update this schema and use it for rendering.

Reviewed By: muirdm

Differential Revision: D44688584

fbshipit-source-id: 286b52990266848fce752d980a2849b7fffc60e4