Add component to change file path display type
Summary:
Add a component to let you change your current display method for files in the uncommitted changes file list.

This component is backed by an atom which itself is backed by an sl config. This means if you reload the page, it will remember what you changed the display type to last time.

I found one bug in the `persistAtomToConfigEffect` code while doing this, it was not distinguishing between different config IDs so they would trample each other.

Reviewed By: quark-zju

Differential Revision: D44774139

fbshipit-source-id: 0f77baef78189d34c9ea4b29589ffc61056d4d83