Add util to extract tree structure from list of paths
Summary:
In the next diffs, I'll add an option to render uncommitted changes as a tree instead of as a linear list

To do this, we'll extract the tree structure from the list of file paths, so we can render as a tree in the DOM.

Here I add a very simple algorithm to construct a map of maps recursively to represent each layer in the tree, to hold an arbitrary data structure per file.

I believe we can assume `/` as the pathsep here, even on windows—though it's probably worth confirming that's what we get from `sl status` since I don't remember and haven't tested windows in a while.

Reviewed By: quark-zju

Differential Revision: D44774138

fbshipit-source-id: 20217d02904e89db0cc535d3b53c3315ccc0607f