Add useDeepMemo
Summary:
In the previous diffs we added a function to build a tree structure for rendering files. This build process constructs a lot of temporary Maps, so it would be ideal to memoize it for performance. However, we need it to re-render if the list of paths changes in any way. This means regular react useMemo won't work, we need deep memoization.

One way to do this might be to use Recoil selectors more aggressively, so we memoize higher in the chain. But we call Uncommitted changes with different parameters in different places simultaneously, so it's a bit more convenient in our rendering to use useMemo instead.

Reviewed By: quark-zju

Differential Revision: D44775134

fbshipit-source-id: f4be2b01126d4bb21ad7758184914a0a16ed0d9c