repo_factory: construct sql more consistently
Summary:
In the repo factory, we do different things for sql connection creation in tests and prod.  In prod we cache `MetadataSqlFactory` instances for each metadata database configuration, and request a new connection from the cached factory each time.  In tests, this doesn't work, as requesting a new sqlite connection from the factory is the same as opening the file again, which can be slow for large numbers of connections, so instead we cache connections directly in the repo factory, which isn't really the right place.

We can fix this by making the `MetadataSqlFactory` handle the caching of sqlite connections for us.  This makes the repo factory consistent between prod and test.  It also means the `MetadataSqlFactory` can hold on to the original connection to the sqlite database and perform the creation queries directly, rather than having to hold it separately in the wrapper around `SqlConnections`, and we can deduplicate the code for opening the database, too.

Reviewed By: mitrandir77

Differential Revision: D44665118

fbshipit-source-id: 1c7d45ac1900d5f2b36285e01159924d0cdb09eb