fsmonitor: fix empty path in dirfilter
## Summary

Running the ignore matcher on an empty path produces incorrect results.
In the rust implementation, this isn't an issue, because there are
directory matchers that run on the directory result. In the Python
treestate bindings, we don't have that functionality, so it returns an
empty string.

This makes this logic match the code in `dirstate.py` (line ~925) where
we check for path being an empty string.

This fixes `sl status` (and related functionality) when fsmonitor is
enabled and there's a complex gitignore file that ignores files in the
root directory.

## Test Plan

Added Mercurial integration test