copytrace: select the last when one file are copied and renamed to multiple files
Summary: We don't have plan to support this one-to-many mapping, since it will make copytrace complexity increase exponentially in theory. For now, we just pick the last copy in ascending alphabetical order. The reason why we do this is to make the selection deterministic, so the behavior will be consistent on different platforms and Rust implementations.

Reviewed By: quark-zju

Differential Revision: D44901629

fbshipit-source-id: 154b515c7043d4a5e45e1f2077f6930baf354e81