prevent extra retry in hgimporter
Summary:
This continues my slashing of fetch retries. This diff is 2x (for a total of
-2 now) fetch retry.

The importer prefetches the tree and then gets the tree. both can cause network
fetches, so when the prefetch fails, the get will retry the network request.

If I am guessing again I think the cause was similar to the last diff:
get use to be a local only check, but now its not anymore so we are exposed to
more potential network fetch retries.

It should be safe to remove this retry because we still kill and restart the
importer if the importer has errors, which should cover any successful re-trys.

With this change we are now down to what looks like 3 network request attempts
per request.

Reviewed By: xavierd

Differential Revision: D44526625

fbshipit-source-id: 455903bdc73c9513106c4e048274d5bb08085a38