config: add ConfigExt::must_get
Summary:
must_get is like get_nonempty_opt but returns an error if the value is unset or
empty. This makes it easy to propagate an error if the desired config is not
set.

I also updated a few spots to use must_get and got rid of a superfluous error type.

Reviewed By: quark-zju

Differential Revision: D44857516

fbshipit-source-id: cdecd7a68d4e3d8248cae977f4cdf7b553b6af74