clone: escape slashes in repo name
Summary:
Canonical repo names can/will have slashes in them. We shouldn't convert slashes to directories on disk; just a bad idea.

When choosing the clone destination, use the basename. It could make sense to clone "foo/bar" into "foo_bar" instead of just "bar", but let's maintain the existing/expected behavior for now.

For internal uses, we percent encode the reponame. For example, the hgcache directory for "foo/bar" will be "foo%2Fbar". This is to minimize chance of collision. If we named the cache dir "foo_bar", that would collide unpleasantly with a repo named "foo_bar".

I relaxed reponame inference a bit to allow slashes in URLs like "test:foo/bar". This was purely so I could test with a reponame containing slashes.

A better long term approach is to use the hash of a canonicalized URL, but that is a bigger change.

Reviewed By: quark-zju

Differential Revision: D44857514

fbshipit-source-id: 4fa0b83a0660b81e89fc96b50ecbeb93869dd2b2