EdenAPI: Changes for ContentMetadataV2 and Blake3
Summary: The `ContentMetadataV2` provides the seeded blake3 hash for each created file. This information needs to be passed to `buck` which interacts with `EdenAPI` through `EdenFS`. The relevant endpoints in `EdenAPI` are `trees` and `files` which were modified in the previous diff to support the new hash and are used here.

Reviewed By: muirdm, markbt

Differential Revision: D44010967

fbshipit-source-id: a755f0750ed39719b1f0e6895b7ca81a405fcfed