Getting file metadata implementation
Summary:
**Context**
D37136735 introduced a new type called `ContentMetadataV2`, to include more information about a file. For example, if a file is binary.

We want to modify the file_info endpoint in SCS, so that users can directly query the metadata for a file.

**This diff**
Implement the modified version of file_info endpoint using the ContentMetadataV2

Reviewed By: markbt

Differential Revision: D43537343

fbshipit-source-id: 98554366a725636d4cd6989fb7e2d9b5fe29d5e4