Modify the file_info endpoint to support additional metadata
Summary:
**Context**
D37136735 introduced a new type called `ContentMetadataV2`, to include more information about a file. For example, if a file is binary.

We want to modify the file_info endpoint in SCS, so that users can directly query the metadata for a file.

**This diff**
Modify Thrift data type, and the function signature for the endpoint

Reviewed By: markbt

Differential Revision: D43537344

fbshipit-source-id: 6076703f67923e0fa95676b0db812a75fb6c3650