move operationBeingPreviewed to serverAPIState to fix import cycle
Summary:
Remove another import cycle from an atom used for previews. This is the only atom used in previews, and it kind of makes more sense alongside other similar operation-related atoms in serverAPIState.

Maybe later we'll move all these things to a new file just for everything operation-related, but I think it's ok in here for now.

Reviewed By: sggutier

Differential Revision: D45017251

fbshipit-source-id: 897583f5dc5d3d366f2a53ac0d4970121e7096b2