global tracker module to prevent import cycles
Summary:
We had several long import cycle chains related to analytics & the Internal abstraction layer. A lot of our internal features use tracking, so they are both exported by Internal and also import Internal for analytics.

I think analytics is a sort of a client that a lot of different features will need. It'll be easier if we can import analytics without worrying about cycles. So here I add a way to access the global tracker off the window object, which requires no imports.

We set the global object when we construct the tracker at the top level. But based on import order, if you try to use the window.getTracker() at the top level, it's possible that it hasn't been set yet. So this API is technically nullable. In practice, these callsites don't use tracking at the top level so it shouldn't matter.

We only need to use this when it causes an import cycle. Otherwise, we get better support by just importing directly.

This is a bit annoying, but it's nice to eliminate a whole category of import cycles.

Reviewed By: quark-zju

Differential Revision: D45017257

fbshipit-source-id: c91c6b114ef67f656f845e291c926cab9047588e