introduce portmapper protocol v2
Summary:
The msft NFS client likes to speak portmapper v2. It does not seem to know v4.
Perhaps there is a way to configure that, but I have not found it yet.

v4 is the latest version of the protocol and so what we should prefer using on
mac and linux.

Here I am introducing the v2 protocol and renaming all the v4 types to
explicitly reference v4.

Reviewed By: chadaustin

Differential Revision: D44987848

fbshipit-source-id: 97f9e2365736b49a5888af204609598b2566de55