remove an unnecessary use of makeSemiFutureWith
Summary:
We overuse make*FutureWith. As a rule, leaf functions should not need
it, because callers must already consider the possibility that an
exception is thrown prior to the asynchronous function's first
`await`. (The same rules apply to Rust async functions, C++
coroutines, Future chains, and Python async.)

Remove an unnecessary makeSemiFutureWith in performBindMounts..

Reviewed By: kmancini

Differential Revision: D44932401

fbshipit-source-id: 5a2a2773cc92faa59a981ee7b1e01324b5d3d5d5