fix startup status underflow
Summary:
When a user has more than 8 mounts we summarize the output of eden start. But
we have an underflow in the number of "finished" mounts.

The issue is that we do:
"number of successfully mounted mounts" - (
    "number of mounts that have not yet started that were already printed" +
    "number of mounts that failed that were already printed" +
    "number of mounts that successfully mounted that were already printed")
and stick that in an unsigned int.

When there are not yet started mounts or failed mounts this number underflows.

This makes the output confusing.

This diff introduces a new category for failed mounts and displays this in the
summary.

And not yet starting mounts are considered "in progress".

With these two fixes that number that unerflowed is now just:
"number of successfully mounted mounts" -
"number of mounts that successfully mounted that were already printed"\

which should not underflow.

Additionally, while testing I noticed we have some output issues with the
errors we display when a mount fails. So I fixed the ordering of our output.

Reviewed By: chadaustin

Differential Revision: D45128100

fbshipit-source-id: 181adaaa7c9dbaba367cb6793ecf648515f4d511