copytrace: make find_renames return early when found the target rename
Summary: The original `find_renames` finds all the renames of the commit and checks if it contains the path that we are interested in. This diff optimizes that logic to make it return early when found the target rename. In order to do that we split `find_renames` into two functions `find_rename_forward` and `find_rename_backward`, since they will have different optimization logics.

Reviewed By: quark-zju

Differential Revision: D45149506

fbshipit-source-id: b77b3b9ef4700581e6d044644d40bf405d66d263