api for vscode config from webview
Summary: In the previous diff, we added the UI to render the opt out / opt  in button. Here I add an API for the vscode webview to read/subscribe to a vscode config and also set the config. This lets us easily implement the opt out flow without needing a special opt out message to be sent to the platform.

Reviewed By: quark-zju

Differential Revision: D45123610

fbshipit-source-id: 5c3ab94960f81becb95c90c9cd27dda3ed8fc1a5