Add command to close ISL webview
Summary: To support opting out of the new ISL, we should close the new ISL when you click the opt out button. This means we need a command to close the webview. We could handle this other ways, but a command is pretty easy to do and useful. The command is not exposed in the package.json so it won't be user-visible.

Reviewed By: quark-zju

Differential Revision: D45129282

fbshipit-source-id: 9feb79701ccd48512a1ddfba89427e72e621c261