stack: initial commitStackState
Summary:
The commitStackState maintains a stack of commits and is intended to provide
read and write operations.

This diff adds basic read operations (cat, log). More complex operations
will be added later.

Right now, the implementation supports forks in the stack (but not merges or
disconnected stacks). If that adds too much undesired complexity we might want
to restrict it to linear stack later.

Reviewed By: evangrayk

Differential Revision: D44845639

fbshipit-source-id: 2a9833dd6fabe7b5c4ecd3f199a362d25f306bd3