stack: support switching file content source of truth between linelog and string
Summary:
Previously, the file content in commit stack is a string, or a base85 string.
This diff extends it to support a 3rd type: lazily defined by (file stack
index, file stack rev). So if we only edit things without showing the file
contents, it might do less work under the hood.

Reviewed By: muirdm

Differential Revision: D45073072

fbshipit-source-id: 1eb351c03a464c7cf3f55af656167dab115f421a