use EventBaseState in TakeoverServer
Summary:
Out of curiosity, I'd removed a .via(getEventBase()) from
TakeoverServer, and it manifested as hard-to-understand segmentation
faults in the integration tests. The issue is that FutureUnixSocket
has EventBase affinity and must only be accessed from it.

To make this safer and more obvious, put FutureUnixSocket and the
other takeover connection state in EventBaseState.

Reviewed By: kmancini

Differential Revision: D44945469

fbshipit-source-id: add3af84a35d451385a73afb622f9b63c7cdb217