scs_server: add support for preparing skeleton manifests
Summary:
Add support for skeleton manifests as a type of derived data that can be prepared.

The current enum for derived data types does not follow best practices and uses zero.  Fix this now by adding a new variant for `FSNODE`.  Once server support is available, we will remove `FSNODE_NEW` and switch to that value for `FSNODE`.

Differential Revision: D45216983

fbshipit-source-id: b0714ef6f9121139cc8d4602a5a8d1beaac00a3f