fix 'repo changed while backing up'
Summary:
fix 'repo changed while backing up'

the current logic is still incorrect for non besteffort (regular) mode

we introduced a high level transaction in D44664729, so now it is safe to drop
the check for non besteffort (regular) mode entirely

Why is it incorrect?

This is because the check is inside the retry loop (have lost the race),
inside the loop we apply commit cloud changes from the currect sync,
so basically if we haven't synced yet at this stage we will always fail the check.

We are loosing our retry logic that we should keep for the regular mode.

After a high level transaction has been added, we can drop the check and
resuscitate our retry logic.

The only possibility for another transaction to change the repo is that the transaction belongs to the previous iteration of the loop in the current cloud sync.

Reviewed By: quark-zju

Differential Revision: D45271716

fbshipit-source-id: 78de00c512f6ac737b107053563bd74b43080d32