Getting started skeleton
Summary:
Scaffolding to support a getting started modal. This modal should only show once on startup, and never again. This means we need to back this state via `sl config`, then check that state on startup.

By lazy loading the modal, the bundle and startup time won't pay the cost of this experience unless its the first time and you need to see it.

The actual experience will be handled by each platform. This is because we currently only want this experience on internal builds and only in our vscode extension. The scaffolding here allows us to give each platform their own experience if they choose, and OSS could get their own too.

Reviewed By: quark-zju

Differential Revision: D45205054

fbshipit-source-id: c97ac87fee1c447535ceb94b8f3f3c2a2cf86dbd