prefetch expensive bookmarks separately
Summary:
prefetch expensive bookmarks separately

this is to avoid timeouts until we have streaming support and ublock rollout
for AOSP repo

also, this provides better ux since we notify users that slow fetch is being executed

I also noticed that the commitcloud.remotebookmarkssync config option wasn't mentioned in the help. So, add it to avoid confusion.

Reviewed By: quark-zju, zzl0

Differential Revision: D44998087

fbshipit-source-id: f24bd959f02380ff0d7ff4f30dbe7e1647403329