translate_id: add support for xrepo lookups
Summary:
Respect the optional "from" and/or "to" repo names to convert the commit ids from and/or to different repos. Only applies for repos that are megarepo'd together.

I added a get_repo method to the EdenApiContext object to facilitate the xrepo behavior.

I thought about disallowing requests where both "from" and "to" are different from the "subject" repo in the path, but it seems harmless.

Reviewed By: quark-zju

Differential Revision: D45019173

fbshipit-source-id: 8a25f5506521765296142553963b19f47570e865