Fix incorrect comparison bases for committed-type comparisons
Summary:
If you click the diff button on a changed file in a commit, it attempts to do a "committed"-type comparison, which is for a specific hash. The hash being used was correct, but the issue was that the `vscode.diff` we opened used the current file contents as the right side, instead of using whatever came before the commit being compared.

instead, to see the diff for hash AAAA, we need to compare left: AAAA with right AAAA^ (the parent of AAAA)

This bug manifested as though the comparison was empty if you viewed the changes for the most recent commit

Reviewed By: quark-zju

Differential Revision: D45331054

fbshipit-source-id: 23361ad4c63cd1591d756a949a9a44a9701a26cc