Upload raw git tags during repo import
Summary: Similar to direct `gitimport`, this diff adds support in `remote-gitimport` to upload raw git annotated tags when pushed to the remote git server. Note that this only uploads annotated tags since basic tags get directly converted into branches while annotated tags need a corresponding changeset to be represented in the Mononoke data model. This diff doesn't delete existing tags when deletions are pushed to the server and modifications of the raw git tag are prevented by the no-update model in Mononoke blobstores.

Reviewed By: markbt

Differential Revision: D45271703

fbshipit-source-id: f7cdbed7fb665c9d32b08e26d7e9f404175b6eb9