Hide uncommit button on closed diffs
Summary:
While legal, it doesn't make a lot of sense to uncommit a Closed Diff / PR. That code has landed, so you usally don't want to be replacing the file state, you'd get weird conflicts if you moved to main.

So let's hide this button to simplify the UI when on a landed commit.

We use the previous diff's API to tell if a diff has landed

Reviewed By: quark-zju

Differential Revision: D45337938

fbshipit-source-id: 3a04dc7104e03856c3802ac6c0fd0ab9c4f74d38