join expensive remote bookmarks into a single fetch
Summary:
join expensive remote bookmarks fetching into a single fetch in cloud sync

This can be done because streaming has been implemented for public pulls!

And if we join it will be faster overall to pull.

Although, this is very rare, I've seen a workspace with oculus-10, oculus-12, and portal-pie branches together.

Reviewed By: mitrandir77

Differential Revision: D45394191

fbshipit-source-id: 901370f4be8b846b1124894554f17cf8b0067d1e