Don't use antialiased font-smoothing
Summary:
The `-webkit-font-smoothing: antialiased` style seems to have been added as part of create-react-app or something. This causes the font to be rendered slightly thinner than the old ISL, making things a bit harder to read according to several users, even though the font-size and color are actually the exact same as the old ISL.

Simply removing this makes our commit titles pop a bit more which makes it a little easier to read.

Reviewed By: cs01

Differential Revision: D45410555

fbshipit-source-id: 8b1e780fb3cd3ae56b646b4cc760d408eed9f8b2