Fix tooltips that rerender not detecting onMouseLeave
Summary:
Kind of a subtle issue here. If you hover on the Pull button long enough to see the tooltip, then click pull, then wait for the tooltip content to change (add "Pull is currently running.") it will no longer detect onMouseLeave and so the tooltip will not auto-dismiss anymore.

This seems to be a quirk of the browser, perhaps related to react synthetic events...? Kind of difficult to attribute exactly, but other people see the same kind of issue: https://github.com/facebook/react/issues/4492

It seems if we add our listener in a layout effect instead of adding it directly via react `onMouseLeave` prop passed ot the div, it keeps the same listener on the DOM node and that means the mouseLeave continues being detected.

Reviewed By: quark-zju

Differential Revision: D45418402

fbshipit-source-id: a1fe0d063f74f4559a75b68177d4828d0f7d8d80