treestate: add fastpath to normalized_get()
Summary: Before we do a slower search for all matching paths, do an optimistic "get()" for the exact path. If it exists and is EXIST_NEXT, we can confidently return it early. I think this is the normal case in practice because case insensitive filesystems are case preserving, so things normally match up between dirstate and filesystem.

Reviewed By: quark-zju

Differential Revision: D45378856

fbshipit-source-id: 7fb5fdadd417d95a39bb559d703dc792e336bca0