blame: use edenapi with "-r wdir()" for clean files
Summary:
Now if the user requests a blame including working copy changes, we still use EdenAPI as long as the file is clean.

For a real fix we can use the server data as a base and then augment it as required. That is non-trivial, so let's do that later.

Note that I noticed we are displaying the wrong line number in blame output. Will fix in following diff.

Reviewed By: quark-zju

Differential Revision: D45415546

fbshipit-source-id: 47f412c835d20107ad945dc030cb8b3eb541bf6e