Log file content type for Metadata Diffs in SCS
Summary: Phabricator uses the `commit_file_diffs` method to get metadata diff which it then uses to render information on the diff view page. With `ContentMetadataV2`, there are opportunities to avoid loading the  files in the commit in memory cause we already have the necessary metadata in hand. However, this optimization will only come in play when the file is non-textual (i.e. binary or non-UTF8). This diff adds logging changes to identify such requests from clients.

Reviewed By: YousefSalama

Differential Revision: D45395092

fbshipit-source-id: 19d2e7370adafff6e626387e1480249349da9be1